refactor: 빌드를 위해 pages안에 스타일들 style 폴더안으로 이동 및 사용하지 않는 스타일들 삭제
