+++ src/styles/registerStyles.tsx
